@@ -1,50 +1,53 @@
 {
   "name": "data-unit-converter",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@primer/octicons-react": "^16.2.0",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^16.11.19",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "gh-pages": "^3.2.3",
-    "mathjs": "^10.0.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "react-select": "^5.2.2",
-    "typescript": "^4.5.4",
-    "web-vitals": "^2.1.3"
-  },
+  "description": "A simple data unit converter.",
+  "version": "1.0.0",
   "scripts": {
+    "start": "webpack serve --config webpack.config.dev.js",
+    "build": "webpack --config webpack.config.prod.js",
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "deploy": "gh-pages -d dist"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+  "keywords": [
+    "react",
+    "data-conversion",
+    "mathjs",
+    "typescript",
+    "webpack",
+    "docker-container"
+  ],
+  "author": "Jhordy Gavinchu <dev@jhordyess.com>",
+  "license": "MIT",
+  "homepage": "https://jhordyess.github.io/data-unit-converter",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jhordyess/data-unit-converter.git"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "bugs": {
+    "url": "https://github.com/jhordyess/data-unit-converter/issues"
   },
-  "homepage": "https://jhordyess.github.io/data-unit-converter"
+  "dependencies": {
+    "@primer/octicons-react": "^17.9.0",
+    "mathjs": "^11.3.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-select": "^5.6.0"
+  },
+  "devDependencies": {
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
+    "clean-webpack-plugin": "^4.0.0",
+    "css-loader": "^6.7.1",
+    "css-minimizer-webpack-plugin": "^4.2.2",
+    "gh-pages": "^4.0.0",
+    "html-webpack-plugin": "^5.5.0",
+    "mini-css-extract-plugin": "^2.6.1",
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.3.6",
+    "ts-loader": "^9.4.1",
+    "typescript": "^4.8.4",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.11.1"
+  }
 }
