@@ -1,5 +1,5 @@
-import React from "react";
-import "./Input.css";
+import * as React from "react";
+import "./style.css";
 const Input = React.forwardRef(
   (
     prop: {
